tutos/requirements.txt
asgiref==3.2.7
astroid==2.3.3
backcall==0.1.0
decorator==4.4.2
Django==3.0.7
django-cors-headers==3.3.0
django-filter==2.2.0
django-guardian==2.2.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
ipython==7.13.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.0
lazy-object-proxy==1.4.3
Markdown==3.2.1
mccabe==0.6.1
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
pkg-resources==0.0.0
prompt-toolkit==3.0.5
psycopg2==2.8.5
ptyprocess==0.6.0
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.4.4
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
traitlets==4.3.3
typed-ast==1.4.1
wcwidth==0.1.9
wrapt==1.11.2
